+++ lib/landable/liquid/tags.rb
@@ -71,21 +71,28 @@ module Landable
       def render(context)
         template = Landable::Template.find_by_slug @template_slug
 
-        if template
-          if template.partial?
-            if context.registers[:responder].present? 
-              context.registers[:responder].controller.render_to_string(partial: template.file)
-            else
-              "<!-- render error: unable to render \"#{@template_slug}\", no controller/responder present -->"
-            end
+        # Handle Templates that are Partials
+        if template && template.partial?
+          responder = context.registers[:responder]
+          # the controller we need for rendering. the request we need to dodge a bug in rails 4.1
+          # (fixed in https://github.com/rails/rails/commit/f6d9b689977c1dca1ed7f149f704d1b4344cd691).
+          # if we need to render pages offline (i.e. without a request) we'll need to rethink this.
+          if responder.try(:controller).try(:request).present?
+            responder.controller.render_to_string(partial: template.file)
           else
-            ::Liquid::Template.parse(template.body).render @variables
+            "<!-- render error: unable to render \"#{@template_slug}\", no controller/responder present -->"
           end
+
+        # Handle Published Templates
+        elsif template && template.revisions.where(is_published: true).present?
+          template = template.revisions.where(is_published: true).first
+          ::Liquid::Template.parse(template.body).render @variables
+
+        # Handle Template Errors
         else
-          "<!-- render error: missing template \"#{@template_slug}\" -->"
+          "<!-- render error: missing published template \"#{@template_slug}\" -->"
         end
       end
     end
-
   end
 end