+++ spec/factories/template_revision.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :template_revision, class: 'Landable::TemplateRevision' do
+    association :template, strategy: :build
+    association :author, strategy: :build
+  end
+end