+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 See README.md before updating this file.
 
-## Unreleased [#]
+## Unreleased [#](https://github.com/enova/landable/compare/v1.7.0...master)
+* Allow for revisions for Templates [#3]